[fix] 고용 버그 수정

+++ Assets/SDH/Scripts/Shop/ShopArtifactItem.cs
@@ -16,7 +16,7 @@ public class ShopArtifactItm : ShopItem
         }
         while (Managers.Artifact.Artifacts[artifactIdx]);
 
-        Managers.Artifact.Artifacts[(int)artifactIdx] = true;
+        Managers.Artifact.Artifacts[artifactIdx] = true;
         artifactTxt.text = ((EArtifacts)artifactIdx).ToString() + "\n100¿ø";
     }
